tests/justhtml-tests/xml_coercion.dat

#data
#xml-coercion
<!-- normal comment -->
#errors
(1,23): expected-doctype-but-got-eof

#document
| <!--  normal comment  -->
| <html>
|   <head>
|   <body>


#data
#xml-coercion
<div>normal text</div>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <div>
|       "normal text"


#data
#xml-coercion
<div>über</div>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <div>
|       "über"


#data
#xml-coercion
<div>über</div>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <div>
|       "über "

#data
<style>a\x0d\x0ab\x0cc\x00</style>
#errors
#xml-coercion
#document
| <html>
|   <head>
|     <style>
|       "a
b c�"
|   <body>

#data
<plaintext>a\x0d\x0ab\x0cc\x00
#errors
#xml-coercion
#document
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "a
b c�"
